service/requirements.txt
tensorflow[and-cuda]
numpy
pandas
requests
pillow
scikit-learn
keras-tuner
opencv-python
flask
jsonify
requests
gunicorn


